+++ cloudfront/invalidation/lambda/src/cacheInvalidation.ts
